xbase integration: remove old ModelState class duplicate
Plus introduce some stubs in new ModelState class.

issue: aurusov/raox#288

+++ ru.bmstu.rk9.rao.lib/src/ru/bmstu/rk9/rao/lib/sequence/Triangular.java
@@ -38,7 +38,7 @@ public class Triangular implements NumericSequence {
 
 	public Double next(double a, double b, double c) {
 		double next = mersenneTwister.nextDouble();
-		double edge = (double) (c - a) / (double) (b - a);
+		double edge = (c - a) / (b - a);
 
 		if (next < edge)
 			return a + Math.sqrt((b - a) * (c - a) * next);
